// prisma/generated/schemas/objects/FormationCountOrderByAggregateInput.schema.ts
import { z } from 'zod';
import { SortOrderSchema } from '../enums/SortOrder.schema';

import type { Prisma } from '@prisma/client';

const Schema: z.ZodType<Prisma.FormationCountOrderByAggregateInput> = z
  .object({
    id: z.lazy(() => SortOrderSchema).optional(),
    createdAt: z.lazy(() => SortOrderSchema).optional(),
    duration: z.lazy(() => SortOrderSchema).optional(),
    location: z.lazy(() => SortOrderSchema).optional(),
    title: z.lazy(() => SortOrderSchema).optional(),
    description: z.lazy(() => SortOrderSchema).optional(),
    image: z.lazy(() => SortOrderSchema).optional(),
    technologies: z.lazy(() => SortOrderSchema).optional(),
    extraLinks: z.lazy(() => SortOrderSchema).optional(),
    speciality: z.lazy(() => SortOrderSchema).optional(),
    landingDescription: z.lazy(() => SortOrderSchema).optional(),
  })
  .strict();

export const FormationCountOrderByAggregateInputObjectSchema = Schema;